+++ auth_backend/auth_method/method_mixins.py
@@ -1,5 +1,12 @@
 from abc import ABCMeta, abstractmethod
 
+from sqlalchemy.orm import Session as DbSession
+
+from auth_backend.auth_method.session import Session
+from auth_backend.models.db import User
+from auth_backend.schemas.types.scopes import Scope as TypeScope
+from auth_backend.utils.user_session_control import create_session
+
 from .base import AuthPluginMeta
 from .session import Session
 
@@ -19,6 +26,14 @@ class RegistrableMixin(AuthPluginMeta, metaclass=ABCMeta):
     async def _register(*args, **kwargs) -> object:
         raise NotImplementedError()
 
+    @staticmethod
+    async def _create_user(*, db_session: DbSession) -> User:
+        """Создает пользователя"""
+        user = User()
+        db_session.add(user)
+        db_session.flush()
+        return user
+
 
 class LoginableMixin(AuthPluginMeta, metaclass=ABCMeta):
     """Сообщает что AuthMethod поддерживает вход
@@ -34,3 +49,10 @@ class LoginableMixin(AuthPluginMeta, metaclass=ABCMeta):
     @abstractmethod
     async def _login(*args, **kwargs) -> Session:
         raise NotImplementedError()
+
+    @staticmethod
+    async def _create_session(
+        user: User, scopes_list_names: list[TypeScope] | None, session_name: str | None = None, *, db_session: DbSession
+    ) -> Session:
+        """Создает сессию пользователя"""
+        return await create_session(user, scopes_list_names, db_session=db_session, session_name=session_name)
