+++ auth_backend/auth_method/oauth.py
@@ -68,17 +68,6 @@ class OauthMeta(UserdataMixin, LoginableMixin, RegistrableMixin, AuthPluginMeta)
             return auth_method.user
 
     @classmethod
-    async def _register_auth_method(cls, key: str, value: str | int, user: User, *, db_session) -> AuthMethod:
-        """Добавление пользователю новый AuthMethod"""
-        return AuthMethod.create(
-            user_id=user.id,
-            auth_method=cls.get_name(),
-            param=key,
-            value=str(value),
-            session=db_session,
-        )
-
-    @classmethod
     async def _delete_auth_methods(cls, user: User, *, db_session) -> list[AuthMethod]:
         """Удаляет пользователю все AuthMethod конкретной авторизации"""
         auth_methods: list[AuthMethod] = (
